Feature/form crud (#121)
* emoji picker and comment form CRUD functionality

* deleted files

* removed console logs

* refactoring....

* FIX: display correct data in post by id page for profile

* only author of comment can edit it

* delete button restricted & type errors fixed

* updated DB messaging

* removed console log

* user can edit and delete their posts

* added contentType to schema

* lexical editor correctly display HTML when editing

* review changes made

* removed commented-out code

* FIX: type error

* Refactored updatePost to edit tags

* FIX: type error

* added emoji picker to lexical editor

* create post form refactor

* FIX: type errors

* refactoring

* refactoring

* resolved conflict

* review changes made

* making sure types are consistent

* dymanic validation for CreatePost

* extracted function from JSX

* added conditional for tags

* feature/upload-podcast

* added additional fields to CreatePost form

* styling inputs

* refactoring

* added functionality to add meetup

* typed action functions

* FIX: type errors in action functions

* deleted un-used code

* made fields nullish for validation

* FIX: type errors

* refactored CreatePostTags for regisrtation input using register

* styling creatable

* refactor with Mateo

* removed router.push

* lexical editor displaying edit text properly - code was removed in commit
4fc2c23 ??????

* Refactor - CSS styles in form

* styled lexical editor

* refactored fetch into single DB call

* removed console logs

* clerkId and userId handled server side

* changed variable name to plural in naming an array

* all create operations are verified on server side

* review changes made

* created CreatePostContext - although unusable, since CreatePost is a server component

* renamed edit/delete popover, moved file location

* commit prior to seeding DB

* updated seed to include clerkId for meetups, interview and podcast

* logged in user can delete their own podcasts

* working on editing podcast

* podcast CRUD

* implemented context API for CreatePost component

* removed redundant useEffect in LexicalMenu

* FIX: hydration error

* user can create, update and delete a meetup

* user can create, edit and delete an interview

* implementing userCanEditMedia

* user can do all CRUD on meetup, interview, post, podcast

* FIX: tags, and banner image BUG

* normalised PodcastWithShowType

* refactored CreatePost - extracted functions into utility.ts

* CRUD error free

* implemeted useContext for displaying previews

* added loading indicator

* set up Toasts for deleting media

* removed console log

* updated seed to include contentType, to align with schema

+++ lib/actions/podcast.actions.ts
@@ -2,14 +2,16 @@
 
 import { type Podcast } from "@prisma/client";
 import prisma from "@/lib/prisma";
+import { redirect } from "next/navigation";
+import { verifyAuth } from "../auth";
 
 type CreatePodcastType = {
+  title: string;
   details: string;
   image: string;
-  showId: number;
-  title: string;
   url: string;
-  userId: number;
+  showId: number;
+  contentType: string;
 };
 
 export async function getPodcastById(podcastId: number) {
@@ -146,13 +148,18 @@ export async function getFilterPodcastsUserInfo({
   }
 }
 
-export async function createPodcast(data: CreatePodcastType) {
+export async function createPodcast(data: CreatePodcastType): Promise<Podcast> {
   try {
-    const podcast = await prisma.podcast.create({
-      data,
-    });
+    const user = verifyAuth("You must be logged in to create post.");
+    const dbUserID: number = (user.sessionClaims.metadata as any).userId;
+    const clerkId: string = user.userId;
 
-    return podcast;
+    if (!dbUserID) throw new Error("User not found");
+
+    await prisma.podcast.create({
+      data: { ...data, userId: dbUserID, clerkId },
+    });
+    redirect("/podcasts");
   } catch (error) {
     console.error("Error creating podcast:", error);
     throw error;
@@ -161,14 +168,22 @@ export async function createPodcast(data: CreatePodcastType) {
 
 export async function updatePodcast(id: number, data: Partial<Podcast>) {
   try {
-    const podcast = await prisma.podcast.update({
+    const user = verifyAuth(
+      "You must be logged in to delete a podcast, and you can only delete yor own podcasts."
+    );
+
+    const dbUserID: number = (user.sessionClaims.metadata as any).userId;
+
+    if (!dbUserID) throw new Error("User not found");
+
+    await prisma.podcast.update({
       where: {
         id,
+        userId: dbUserID,
       },
       data,
     });
-
-    return podcast;
+    redirect("/podcasts");
   } catch (error) {
     console.error("Error updating podcast:", error);
     throw error;
@@ -177,13 +192,23 @@ export async function updatePodcast(id: number, data: Partial<Podcast>) {
 
 export async function deletePodcast(id: number) {
   try {
-    const deletedPodcast = await prisma.podcast.delete({
+    const user = verifyAuth(
+      "You must be logged in to delete a podcast, and you can only delete yor own podcasts."
+    );
+
+    const dbUserID: number = (user.sessionClaims.metadata as any).userId;
+    const clerkId: string = user.userId;
+
+    if (!dbUserID) throw new Error("User not found");
+
+    await prisma.podcast.delete({
       where: {
         id,
+        clerkId,
       },
     });
 
-    return deletedPodcast;
+    redirect("/podcasts");
   } catch (error) {
     console.error("Error deleting podcast:", error);
     throw error;
@@ -234,6 +259,81 @@ export async function getTopFiveShowIds() {
     return showIds;
   } catch (error) {
     console.error("Error fetching the top five show IDs:", error);
+    throw error;
+  }
+}
+
+interface ShowOption {
+  label: string;
+  value: number;
+}
+
+export async function getAllShowOptions(): Promise<ShowOption[]> {
+  try {
+    const shows = await prisma.shows.findMany({
+      select: {
+        id: true,
+        name: true,
+      },
+    });
+
+    const showOptions: ShowOption[] = shows.map((show) => ({
+      label: show.name,
+      value: show.id,
+    }));
+
+    return showOptions;
+  } catch (error) {
+    console.error("Error fetching all shows:", error);
+    throw error;
+  }
+}
+
+export type PodcastWithShow = {
+  heading: string;
+  content: string;
+  image: string;
+  podcast: string;
+  contentType: string;
+  show: { label: string; value: string };
+};
+
+export async function getPodcastToEditById(
+  mediaId: number
+): Promise<PodcastWithShow | null> {
+  try {
+    const user = verifyAuth(
+      "You must be logged in to delete a podcast, and you can only delete your own podcasts."
+    );
+    const dbUserID: number = (user.sessionClaims.metadata as any).userId;
+    const clerkId: string = user.userId;
+
+    if (!dbUserID) throw new Error("User not found");
+
+    const podcast = await prisma.podcast.findUnique({
+      where: {
+        id: mediaId,
+        clerkId,
+      },
+      include: { show: true },
+    });
+
+    if (!podcast) {
+      return null;
+    }
+
+    const podcastNormalised = {
+      heading: podcast.title,
+      content: podcast.details,
+      image: podcast.image,
+      podcast: podcast.url,
+      contentType: podcast.contentType,
+      show: { label: podcast.show.name ?? "", value: String(podcast.show.id) },
+    };
+
+    return podcastNormalised;
+  } catch (error) {
+    console.error("Error fetching podcast by ID:", error);
     throw error;
   }
 }
